Update Dockerfile
Aravis 0.8 can now be installed via apt. This reduces a large amount of dependencies and manual steps.

The first stage is separated in multiple RUN steps to better show
the steps taken. The amount of layers does not matter much as
this stage is just used to create the packaged libraries. This stage
is not used or uploaded.

The libraries are no longer copied, but mounted.