Move init logic for resolution to own function (refactor)
The logic had a bug (comparing it to the serial
instead of to the resolution). Fixing the logic was
more diffucult than it needed to be due to the
single scope and multiple goto statements.

The resolution is now also stored on in the node as
a parameter. No need to have that info multiple
times.